+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## [1.0.0-beta.11] - 03-08-2017
+### Fixed
+- Use of `CHECK_TOKENS` server flag to reduce overhead of pulling token when not needed
+- Change ping to touch
+- Re-estabilish connection if remote disconnected and dispatch all the logs
+- Updated touch threshold to 2 minutes
+
 ## [1.0.0-beta.10]
 ### Fixed
 - Fixed issue with token retrieving when client is dead